requirements.txt
tensorflow-cpu==2.13.0
transformers==4.31.0
scikit-learn==1.3.1
joblib==1.3.0
pandas==2.1.3
streamlit==1.25.0
altair==4.2.2
streamlit-extras==0.5.0
streamlit-lottie